remove server.rs and adjust app
